requirements.txt
altgraph==0.17.4
asttokens==2.4.1
Babel==2.13.1
bqplot==0.12.42
branca==0.7.0
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
colour==0.1.5
comm==0.2.0
contourpy==1.2.0
cycler==0.12.1
decorator==5.1.1
earthengine-api==0.1.378
eerepr==0.0.4
exceptiongroup==1.1.3
executing==2.0.1
folium==0.15.0
fonttools==4.44.0
future==0.18.3
geemap==0.29.5
geocoder==1.38.1
google-api-core==2.14.0
google-api-python-client==2.107.0
google-auth==2.23.4
google-auth-httplib2==0.1.1
google-cloud-core==2.3.3
google-cloud-storage==2.13.0
google-crc32c==1.5.0
google-oauth2-tool==0.0.3
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
httplib2==0.22.0
idna==3.4
ipyevents==2.0.2
ipyfilechooser==0.6.0
ipyleaflet==0.17.4
ipython==8.17.2
ipytree==0.2.2
ipywidgets==8.1.1
jedi==0.19.1
Jinja2==3.1.2
jupyterlab-widgets==3.0.9
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.1
matplotlib-inline==0.1.6
numpy==1.26.1
oauth2client==4.1.3
packaging==23.2
pandas==2.1.3
parso==0.8.3
pefile==2023.2.7
Pillow==10.1.0
plotly==5.18.0
prompt-toolkit==3.0.40
protobuf==4.25.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyCRS==1.0.2
Pygments==2.16.1
pyinstaller==6.1.0
pyinstaller-hooks-contrib==2023.10
pyparsing==3.1.1
pyperclip==1.8.2
pyshp==2.3.1
python-box==7.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
pywin32-ctypes==0.2.2
ratelim==0.1.6
requests==2.31.0
rsa==4.9
scooby==0.9.2
six==1.16.0
stack-data==0.6.3
tenacity==8.2.3
tk==0.1.0
tkcalendar==1.6.1
traitlets==5.13.0
traittypes==0.2.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.7
wcwidth==0.2.9
widgetsnbextension==4.0.9
xyzservices==2023.10.1
